# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KingSaint Capital - Premium Crypto Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /investments:
    get:
      operationId: getInvestments
      tags: [investments]
      summary: Get user investments
      responses:
        "200":
          description: User investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Create new investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "201":
          description: Investment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /investments/{id}:
    get:
      operationId: getInvestment
      tags: [investments]
      summary: Get investment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Investment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get transaction history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal, profit, bonus]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, cancelled]
        - name: limit
          in: query
          schema:
            type: number
            default: 20
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Request deposit or withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, email, fullName, role, status, balance, totalInvested, totalProfit, referralCode, createdAt]
      properties:
        id:
          type: number
        email:
          type: string
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, pending]
        balance:
          type: number
        totalInvested:
          type: number
        totalProfit:
          type: number
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    RegisterInput:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          
        password:
          type: string
          minLength: 8
        fullName:
          type: string
          minLength: 2
        referralCode:
          type: string
        phone:
          type: string
        country:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          
        password:
          type: string
...
    Investment:
      type: object
      required: [id, userId, planId, planName, amount, roiPercent, durationDays, status, startDate, endDate, profitEarned, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        planId:
          type: number
        planName:
          type: string
        planTier:
          type: string
        amount:
          type: number
        roiPercent:
          type: number
        durationDays:
          type: number
        status:
          type: string
          enum: [active, completed, cancelled]
        startDate:
          type: string
        endDate:
          type: string
        profitEarned:
          type: number
        createdAt:
          type: string
    InvestmentInput:
      type: object
      required: [planId, amount, cryptoType, walletAddress]
      properties:
        planId:
          type: number
        amount:
          type: number
        cryptoType:
          type: string
          enum: [BTC, ETH, USDT, BNB, SOL]
        walletAddress:
          type: string
    Transaction:
      type: object
      required: [id, userId, type, amount, status, cryptoType, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        userEmail:
          type: ["string", "null"]
        userFullName:
          type: ["string", "null"]
        type:
          type: string
          enum: [deposit, withdrawal, profit, bonus, referral]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
        cryptoType:
          type: string
        walletAddress:
          type: ["string", "null"]
        txHash:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TransactionInput:
      type: object
      required: [type, amount, cryptoType]
      properties:
        type:
          type: string
          enum: [deposit, withdrawal]
        amount:
          type: number
        cryptoType:
          type: string
          enum: [BTC, ETH, USDT, BNB, SOL]
        walletAddress:
          type: string
        txHash:
          type: string
...