# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * KingSaint Capital - Premium Crypto Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  email: string;
  fullName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
  role: UserRole;
  status: UserStatus;
  balance: number;
  totalInvested: number;
  totalProfit: number;
  referralCode: string;
  /** @nullable */
  referredBy?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}